Split off preparser from JmlIo
