+++ key.core/src/main/java/de/uka/ilkd/key/speclang/jml/JMLSpecExtractor.java
@@ -24,6 +24,7 @@ import de.uka.ilkd.key.speclang.njml.LabeledParserRuleContext;
 import de.uka.ilkd.key.speclang.*;
 import de.uka.ilkd.key.speclang.jml.pretranslation.*;
 import de.uka.ilkd.key.speclang.jml.translation.JMLSpecFactory;
+import de.uka.ilkd.key.speclang.njml.PreParser;
 import de.uka.ilkd.key.speclang.translation.SLTranslationException;
 import de.uka.ilkd.key.speclang.translation.SLWarningException;
 import org.antlr.v4.runtime.ParserRuleContext;
@@ -279,12 +280,10 @@ public final class JMLSpecExtractor implements SpecExtractor {
             Position pos = comments[0].getStartPosition();
 
             // call preparser
-            JmlIO jmlIO = new JmlIO();
-            // TODO wiesler
+            var parser = new PreParser();
             ImmutableList<TextualJMLConstruct> constructs =
-                jmlIO.parseClassLevel(concatenatedComment, fileName, pos);
-
-            warnings = warnings.append(jmlIO.getWarnings());
+                parser.parseClassLevel(concatenatedComment, fileName, pos);
+            warnings = warnings.append(parser.getWarnings());
 
             // create class invs out of textual constructs, add them to result
             for (TextualJMLConstruct c : constructs) {
@@ -357,10 +356,10 @@ public final class JMLSpecExtractor implements SpecExtractor {
             Position pos = comments[0].getStartPosition();
 
             // call preparser
-            JmlIO io = new JmlIO();
+            PreParser parser = new PreParser();
             // TODO wiesler
-            constructs = io.parseClassLevel(concatenatedComment, fileName, pos);
-            warnings = warnings.append(io.getWarnings());
+            constructs = parser.parseClassLevel(concatenatedComment, fileName, pos);
+            warnings = warnings.append(parser.getWarnings());
         } else {
             constructs = ImmutableSLList.<TextualJMLConstruct>nil();
         }
@@ -637,12 +636,10 @@ public final class JMLSpecExtractor implements SpecExtractor {
         }
         final String concatenatedComment = concatenate(comments);
         final Position position = comments[0].getStartPosition();
-        final JmlIO io = new JmlIO();
-        // TODO wiesler
-
+        final var parser = new PreParser();
         final ImmutableList<TextualJMLConstruct> constructs =
-            io.parseMethodLevel(concatenatedComment, fileName, position);
-        warnings = warnings.append(io.getWarnings());
+            parser.parseMethodLevel(concatenatedComment, fileName, position);
+        warnings = warnings.append(parser.getWarnings());
         return constructs.toArray(new TextualJMLConstruct[constructs.size()]);
     }
 
@@ -671,11 +668,10 @@ public final class JMLSpecExtractor implements SpecExtractor {
         Position pos = comments[0].getStartPosition();
 
         // call preparser
-        JmlIO io = new JmlIO();
-        // TODO wiesler
+        var parser = new PreParser();
         ImmutableList<TextualJMLConstruct> constructs =
-            io.parseMethodLevel(concatenatedComment, fileName, pos);
-        warnings = warnings.append(io.getWarnings());
+            parser.parseMethodLevel(concatenatedComment, fileName, pos);
+        warnings = warnings.append(parser.getWarnings());
 
         // create JML loop invariant out of last construct
         if (constructs.size() == 0) {
