+++ key.core/src/test/java/de/uka/ilkd/key/speclang/jml/pretranslation/TextualJMLAssertStatementTest.java
@@ -1,7 +1,7 @@
 package de.uka.ilkd.key.speclang.jml.pretranslation;
 
 import de.uka.ilkd.key.java.Position;
-import de.uka.ilkd.key.speclang.njml.JmlIO;
+import de.uka.ilkd.key.speclang.njml.PreParser;
 import org.junit.jupiter.api.Test;
 import org.key_project.util.collection.ImmutableList;
 
@@ -10,7 +10,7 @@ import static org.junit.jupiter.api.Assertions.*;
 
 public class TextualJMLAssertStatementTest {
     private static ImmutableList<TextualJMLConstruct> parseMethodLevel(String ms) {
-        return new JmlIO().parseMethodLevel(ms, "test", new Position(0, 0));
+        return new PreParser().parseMethodLevel(ms, "test", new Position(0, 0));
     }
 
     private static void assertTextRepr(String input, String text) {
